Deny defaults for higher-ranked generic parameters
