--- tests/ui/parser/issue-119042.rs
@@ -1,7 +1,0 @@
-// check-pass
-
-macro_rules! a { ($ty:ty) => {} }
-
-a! { for<T = &i32> fn() }
-
-fn main() {}
